requirements.txt
cycler==0.11.0
fonttools==4.28.4
h5py==3.6.0
imageio==2.13.3
kiwisolver==1.3.2
matplotlib==3.5.1
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyWavelets==1.2.0
scikit-image==0.17.2
scipy==1.7.3
six==1.16.0
tifffile==2021.11.2
torch==1.10.1+cu102
torchaudio==0.10.1+cu102
torchvision==0.11.2+cu102
typing_extensions==4.0.1
